requirements.txt
disnake
disnake[voice]
PyNaCl
pillow
youtube_dl
googletrans
httpx
httpx[http2]